+++ packages/core/src/agent/index.ts
@@ -517,7 +517,7 @@ export class Agent<
     threadId?: string;
     context?: CoreMessage[];
     runId?: string;
-    messages: UserContent[];
+    messages: CoreMessage[];
   }) {
     return {
       before: async () => {
@@ -528,15 +528,7 @@ export class Agent<
           content: `${this.instructions}. Today's date is ${new Date().toISOString()}`,
         };
 
-        const userMessages: CoreMessage[] = messages.map(content => ({
-          role: 'user',
-          content: content,
-        }));
-
-        let coreMessages = userMessages;
-
-        let convertedTools: Record<string, CoreTool> | undefined;
-
+        let coreMessages = messages;
         let threadIdToUse = threadId;
 
         if (this.#mastra?.memory && resourceid) {
@@ -544,7 +536,7 @@ export class Agent<
             resourceid,
             runId,
             threadId: threadIdToUse,
-            messages: userMessages,
+            messages,
           });
 
           coreMessages = preExecuteResult.coreMessages;
@@ -554,6 +546,8 @@ export class Agent<
             runId: this.name,
           });
         }
+
+        let convertedTools: Record<string, CoreTool> | undefined;
 
         if (
           (toolsets && Object.keys(toolsets || {}).length > 0) ||
@@ -650,17 +644,25 @@ export class Agent<
       output?: OutputType | Z;
     } = {},
   ): Promise<GenerateReturn<Z>> {
-    let messagesToUse: UserContent[] = [];
+    let messagesToUse: CoreMessage[] = [];
 
-    if (Array.isArray(messages)) {
-      messagesToUse = messages.map(content => {
-        if (typeof content === 'string') {
-          return content;
-        }
-        return content.content as string;
-      });
+    if (typeof messages === `string`) {
+      messagesToUse = [
+        {
+          role: 'user',
+          content: messages,
+        },
+      ];
     } else {
-      messagesToUse = [messages];
+      messagesToUse = messages.map(message => {
+        if (typeof message === `string`) {
+          return {
+            role: 'user',
+            content: message,
+          };
+        }
+        return message;
+      });
     }
 
     const { before, after } = this.__primitive({
@@ -739,17 +741,25 @@ export class Agent<
       output?: OutputType | Z;
     } = {},
   ): Promise<StreamReturn<Z>> {
-    let messagesToUse: UserContent[] = [];
+    let messagesToUse: CoreMessage[] = [];
 
-    if (Array.isArray(messages)) {
-      messagesToUse = messages.map(content => {
-        if (typeof content === 'string') {
-          return content;
-        }
-        return content.content as string;
-      });
+    if (typeof messages === `string`) {
+      messagesToUse = [
+        {
+          role: 'user',
+          content: messages,
+        },
+      ];
     } else {
-      messagesToUse = [messages];
+      messagesToUse = messages.map(message => {
+        if (typeof message === `string`) {
+          return {
+            role: 'user',
+            content: message,
+          };
+        }
+        return message;
+      });
     }
 
     const { before, after } = this.__primitive({