+++ examples/bird-checker-with-express/CHANGELOG.md
@@ -1,5 +1,22 @@
 # bird-checker-with-express
 
+## 1.0.1-alpha.48
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/engine@0.0.5-alpha.62
+
+## 1.0.1-alpha.47
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/engine@0.0.5-alpha.61
+
 ## 1.0.1-alpha.46
 
 ### Patch Changes