+++ examples/llm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # mastra-examples-llm
 
+## 1.0.1-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 1.0.1-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 1.0.1-alpha.43
 
 ### Patch Changes