+++ examples/workflow-with-memory/CHANGELOG.md
@@ -1,5 +1,22 @@
 # workflow
 
+## 1.0.1-alpha.27
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/memory@0.0.2-alpha.48
+
+## 1.0.1-alpha.26
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/memory@0.0.2-alpha.47
+
 ## 1.0.1-alpha.25
 
 ### Patch Changes