+++ integrations/stabilityai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mastra/stabilityai
 
+## 1.0.1-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 1.0.1-alpha.42
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 1.0.1-alpha.41
 
 ### Patch Changes