+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,31 @@
 # mastra
 
+## 0.1.57-alpha.105
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+  - @mastra/deployer@0.0.1-alpha.5
+
+## 0.1.57-alpha.104
+
+### Patch Changes
+
+- 7babd5c: CLI build and other
+- Updated dependencies [7babd5c]
+  - @mastra/deployer@0.0.1-alpha.4
+
+## 0.1.57-alpha.103
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+- Updated dependencies [a291824]
+  - @mastra/core@0.1.27-alpha.67
+  - @mastra/deployer@0.0.1-alpha.3
+
 ## 0.1.57-alpha.102
 
 ### Patch Changes