+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mastra/mcp
 
+## 0.0.1-alpha.8
+
+### Patch Changes
+
+- Updated dependencies [0be7181]
+- Updated dependencies [0be7181]
+  - @mastra/core@0.1.27-alpha.68
+
+## 0.0.1-alpha.7
+
+### Patch Changes
+
+- Updated dependencies [c8ff2f5]
+  - @mastra/core@0.1.27-alpha.67
+
 ## 0.0.1-alpha.6
 
 ### Patch Changes